ClassWork/Class_04_10072025/run_streamed/main.py
import os
import asyncio
from agents import Agent, Runner, OpenAIChatCompletionsModel, AsyncOpenAI, set_tracing_disabled
from openai.types.responses import ResponseTextDeltaEvent
from dotenv import load_dotenv

load_dotenv()

set_tracing_disabled(True)

gemini_api_key = os.getenv("GEMINI_API_KEY")
if not gemini_api_key:
    raise ValueError("GEMINI_API_KEY is not set")

MODEL = "gemini-2.0-flash"
BASE_URL = "https://generativelanguage.googleapis.com/v1beta/openai/"

client = AsyncOpenAI(
    api_key=gemini_api_key,
    base_url=BASE_URL,
)

model = OpenAIChatCompletionsModel(
    model=MODEL,
    openai_client=client,
)

agent = Agent(
    name="Assistant",
    instructions= "You are helpful assistant.",
    model=model,
)

async def main():

    result = Runner.run_streamed(
        starting_agent=agent,
        input="Tell me something about Agentic AI in 1000 words"
    )

    # print(result.final_output)

    async for event in result.stream_events():
        if event.type == "raw_response_event" and isinstance (event.data, ResponseTextDeltaEvent):
            print(event.data.delta, end="", flush=True)

if __name__ == "__main__":
    asyncio.run(main())